+++ NEWS.md
@@ -3,7 +3,7 @@
 * Adding 1 hour of downloading to `get_wfs()` for big shape
 * Remove connection to IGN news when library is load. Now `get_last_news()` can be used
 to retrieve last news.
-* Adding `interactive` parameter to `get_wfs` to allow quick use
+* Adding `interactive` parameter to `get_wfs` and `get_wms_raster` to allow quick use
 
 # happign 0.1.6
 * `get_wms_raster()` use gdal with `sf::gdal_utils()` for downloading and `terra::vrt()` for